requirements.txt
Flask==3.0.3
Flask-HTTPAuth==4.8.0
psycopg2-binary==2.9.9
redis==5.0.8
requests==2.32.3
pytest==8.3.2
pytest-flask==1.3.0
python-dotenv==1.0.1
PyJWT==2.9.0

